requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-durationwidget==1.0.5
django-markdown==0.8.4
django-markdownx==3.0.1
gunicorn==20.0.4
Markdown==3.3.4
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
